Run Lua Interpreter

Executes DCS World's luae.exe Lua 5.1.5 interpreter.

luae.exe(...)

Arguments

...

Arguments passed on to base::system2

command

the system command to be invoked, as a character string.

args

a character vector of arguments to command.

stdout,stderr

where output to stdout or stderr should be sent. Possible values are "", to the R console (the default), NULL or FALSE (discard output), TRUE (capture the output in a character vector) or a character string naming a file.

stdin

should input be diverted? "" means the default, alternatively a character string naming a file. Ignored if input is supplied.

input

if a character vector is supplied, this is copied one string per line to a temporary file, and the standard input of command is redirected to the file.

env

character vector of name=value strings to set environment variables.

wait

a logical (not NA) indicating whether the R interpreter should wait for the command to finish, or run it asynchronously. This will be ignored (and the interpreter will always wait) if stdout = TRUE or stderr = TRUE. When running the command asynchronously, no output will be displayed on the Rgui console in Windows (it will be dropped, instead).

timeout

timeout in seconds, ignored if 0. This is a limit for the elapsed time running command in a separate process. Fractions of seconds are ignored.

minimized

logical (not NA), indicates whether the command window should be displayed initially as a minimized window.

invisible

logical (not NA), indicates whether the command window should be visible on the screen.

Details

You do not need to shell-quote standard input, as you might expect. However, you do need to quote for Lua snippets passing through the command line arguments. See examples.

Examples

if (FALSE) {
library(dcsworldr)

luae.exe()
#> Lua 5.1.5  Copyright (C) 1994-2012 Lua.org, PUC-Rio
#> >

luae.exe(input = "print(123)")
luae.exe(input = 'print("hello world")', stdout = TRUE)
luae.exe(c("-e", shQuote('print("hello world")')), stdout = TRUE)
luae.exe(c("-e", "io.write(package.path)"), stdout = TRUE)
luae.exe(c("-e", "io.write(package.cpath)"), stdout = TRUE)

# https://www.lua.org/pil/8.2.html
luae.exe(c("-e",
  shQuote("assert(package.loadlib('lua-socket', 'luaopen_socket_core'))()")))
}